requirements.txt
openai==1.59.9
tavily-python==0.5.0
python-dotenv==1.0.1